+++ client_src/js/lib/modules/printerManager.js
@@ -208,9 +208,407 @@ export default class PrinterManager {
         rotate90 = "rotate(90deg)";
       }
     }
+    let systemSettings = await OctoFarmClient.get("settings/client/get");
 
+    systemSettings = await systemSettings.json();
+    let controlSettings = systemSettings.controlSettings;
     //Load tools
-    document.getElementById("printerControls").innerHTML = `
+    if (typeof controlSettings !== "undefined" && controlSettings.filesTop) {
+      document.getElementById("printerControls").innerHTML = `
+          <div class="row">
+              <div class="col-lg-3">
+             <div class="row">
+                <div class="col-12">
+                <center>
+                <h5>Operation</h5>
+            </center>
+            <hr>
+  
+            <center>
+            <button id="pmPrintStart" type="button" class="btn btn-success" role="button"><i class="fas fa-print"></i> Print</button>
+            <button id="pmPrintPause" type="button" class="btn btn-light" role="button" disabled><i class="fas fa-pause"></i> Pause</button>
+            <button id="pmPrintRestart" type="button" class="btn btn-danger" role="button"><i class="fas fa-undo"></i> Restart</button>
+            <button id="pmPrintResume" type="button" class="btn btn-success" role="button"><i class="fas fa-redo"></i> Resume</button>
+            <button id="pmPrintStop" type="button" class="btn btn-danger" disabled><i class="fas fa-square"></i> Cancel</button>
+            </center></div></div>
+                 <div id="cameraRow" class="row">
+                      <div class="col-12">
+                            <center>
+                                <h5>Camera</h5>
+                            </center>
+                            <hr>
+                        </div>
+                    </div>
+                    <div class="row">
+                        <div id="cameraCol" class="col-12">
+                          <img style="transform: ${flipH} ${flipV} ${rotate90};" id="printerControlCamera" width="100%" src=""/>
+                        </div>
+                    </div>
+                                  <div class="row">
+                    <div class="col-9">
+                        <center>
+                            <h5>X/Y</h5>
+                        </center>
+                        <hr>
+                    </div>
+                    <div class="col-3">
+                        <center>
+                            <h5>Z</h5>
+                        </center>
+                        <hr>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-3"></div>
+                    <div class="col-3">
+                        <center><button id="pcYpos" type="button" class="btn btn-light"><i class="fas fa-arrow-up"></i></button></center>
+                    </div>
+                    <div class="col-3"></div>
+                    <div class="col-3">
+                        <center><button id="pcZpos"type="button" class="btn btn-light"><i class="fas fa-arrow-up"></i></button></center>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-3">
+                        <center><button id="pcXneg" type="button" class="btn btn-light"><i class="fas fa-arrow-left"></i></button></center>
+                    </div>
+                    <div class="col-3">
+                        <center><button id="pcXYhome" type="button" class="btn btn-light"><i class="fas fa-home"></i></button></center>
+                    </div>
+                    <div class="col-3">
+                        <center><button id="pcXpos" type="button" class="btn btn-light"><i class="fas fa-arrow-right"></i></button></center>
+                    </div>
+                    <div class="col-3">
+                        <center><button id="pcZhome" type="button" class="btn btn-light"><i class="fas fa-home"></i></button></center>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-3"></div>
+                    <div class="col-3">
+                        <center><button id="pcYneg" type="button" class="btn btn-light"><i class="fas fa-arrow-down"></i></button></center>
+                    </div>
+                    <div class="col-3"></div>
+                    <div class="col-3">
+                        <center><button id="pcZneg" type="button" class="btn btn-light"><i class="fas fa-arrow-down"></i></button></center>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-12">
+                        <center>
+                            <div id="pcAxisSteps" class="btn-group" role="group">
+                                <button id="pcAxisSteps01" type="button" class="btn btn-light" value="01">0.1</button>
+                                <button id="pcAxisSteps1" type="button" class="btn btn-light" value="1">1</button>
+                                <button id="pcAxisSteps10" type="button" class="btn btn-light" value="10">10</button>
+                                <button id="pcAxisSteps100" type="button" class="btn btn-light" value="100">100</button>
+                            </div>
+                        </center>
+                    </div>
+                </div>
+                              <div class="row">
+                    <div class="col-12">
+                        <center>
+                            <h5>Extruder</h5>
+                        </center>
+                        <hr>
+                    </div>
+                </div>
+                                      <div class="row">
+                    <div class="col-12">
+                        <center>
+                            <div class="input-group">
+                                <input id="pcExtruder" type="number" class="form-control" placeholder="0" aria-label="Recipient's username" aria-describedby="basic-addon2">
+                                <div class="input-group-append">
+                                    <span class="input-group-text" id="basic-addon2">mm</span>
+                                </div>
+                            </div>
+                        </center>
+                    </div>
+                <div class="row">
+                    <div class="col-12 text-center">
+                        <center><button id="pcExtrude" class="btn btn-light" type="submit"><i class="fas fa-redo"></i> Extrude</button> <button id="pcRetract" class="btn btn-light" type="submit"><i class="fas fa-undo"></i> Retract</button></center>
+                </div>
+            </div>
+            </div>
+                  <div class="row">
+                    <div class="col-12">
+                        <center>
+                            <h5>Feed/Flow</h5>
+                        </center>
+                        <hr>
+                    </div>
+                </div>
+                  <div class="row">
+                    <div class="col-10 col-lg-8 col-xl-8">
+                        <label for="pcFeed">Feed Rate: <span id="pcFeedValue">${printer.feedRate}%</span></label>
+                        <input type="range" class="octoRange custom-range" min="50" max="150" step="1" id="pcFeed" value="${printer.feedRate}">
+                    </div>
+                    <div class="col-2 col-lg-4 col-xl-4">
+                        <button id="pcFeedRate" type="button" class="btn btn-light">Update</button>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-10 col-lg-8 col-xl-8">
+                        <label for="pcFlow">Flow Rate: <span id="pcFlowValue">${printer.flowRate}%</span></label>
+                        <input type="range" class="octoRange custom-range" min="75" max="125" step="1" id="pcFlow" value="${printer.flowRate}">
+                    </div>
+                    <div class="col-2 col-lg-4 col-xl-4">
+                        <button id="pcFlowRate" type="button" class="btn btn-light">Update</button>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-12">
+                        <center>
+                            <h5>Motors / Fans</h5>
+                        </center>
+                        <hr>
+                    </div>
+                </div>
+                  <div class="row">
+                    <div class="col-12">
+                        <center><button id="pcMotorTog" class="btn btn-light" type="submit">Motors Off</button></center>
+                    </div>
+                </div>
+                <div class="row">
+                    <div class="col-12">
+                    <label for="pcFlow">Fan Percent: <span id="pcFanPercent">100%</span></label>
+                    <input type="range" class="octoRange custom-range" min="0" max="100" step="1" id="pcFanPercent" value="100">
+                        <center><button id="pcFanOn" class="btn btn-light" type="submit">Set Fans</button> <button id="pcFanOff" class="btn btn-light" type="submit">Fans Off</button></center>
+                    </div>
+                </div>
+              </div>
+              
+              <div class="col-lg-9 pt-0">
+                  <div class="row">
+                    <div class="col-12">
+                        <center>
+                            <h5>Files</h5>
+                        </center>
+                        <hr>
+                    </div>
+                </div>
+                <div class="row bg-secondary rounded-top">
+                <div class="col-12">
+                     <h5 class="float-left  mb-0">
+                      <button id="printerFileCount" type="button" class="btn btn-secondary float-right d-block" href="#" aria-expanded="false" disabled="">
+                        <i class="fas fa-file"></i> Loading... <i class="fas fa-folder"></i> Loading...
+                      </button>
+                      <button id="printerStorage" type="button" class="btn btn-secondary float-right d-block" href="#" aria-expanded="false" disabled="">
+  
+                        <i class="fas fa-hdd"></i> Loading...
+                      </button>
+                    </h5>
+                    <h5 class="float-left mb-0">
+                      <button type="button" class="btn btn-secondary float-right d-block" href="#" aria-expanded="false" disabled="">
+                        <i class="fas fa-file-code"></i> Files: <span id="currentFolder">local</span>/
+                      </button>
+                    </h5>
+                    <div class="btn btn-secondary form-group float-right  mb-0">
+                      <form class="form-inline">
+                        <div class="form-group">
+                          <label for="searchFiles">
+                            <i class="fas fa-search pr-1"></i>
+                          </label>
+                          <input id="searchFiles" type="text" placeholder="File Search..." class="search-control search-control-underlined">
+                        </div>
+                      </form>
+                    </div>
+                   </div>
+                  </div>
+                  <div class="row bg-secondary rounded-bottom">
+                    <div class="col-lg-2">
+                      <i class="fas fa-file-upload ml-2 mb-1"></i><span id="fileCounts-${printer._id}"> 0 </span>
+                    </div>
+                    <div class="col-lg-10">
+                      <div class="progress">
+                        <div id="fileProgress-${printer._id}" class="progress-bar progress-bar-striped bg-warning" role="progressbar" style="width: 0%" aria-valuenow="0" aria-valuemin="0" aria-valuemax="100">
+                          0%
+                        </div>
+                      </div>
+                    </div>
+                  </div>
+                  <div class="row mb-1">
+                      <div class="col-12">
+                       <button id="fileBackBtn" type="button" class="btn btn-success float-right">
+                        <i class="fas fa-chevron-left"></i> Back
+                      </button>
+                      <!-- Split dropright button -->
+                      <div class="float-right mr-3 btn-group">
+                          <div id="fileSortDropdownMenu" class="btn bg-secondary">Sort</div>
+                          <button type="button" class="btn btn-secondary dropdown-toggle dropdown-toggle-split" data-toggle="dropdown" aria-haspopup="true" aria-expanded="false">
+                            <span class="sr-only">Toggle Dropdown</span>
+                          </button>
+                          <div class="dropdown-menu">
+                            
+                       <a class="dropdown-item" id="sortFileNameDown"><i class="fas fa-sort-alpha-down"></i> File Name</a>
+  
+                      <a class="dropdown-item" id="sortFileNameUp"><i class="fas fa-sort-alpha-up"></i> File Name</a>
+                             <div class="dropdown-divider"></div>
+                      <a class="dropdown-item" id="sortPrintTimeDown"><i class="fas fa-sort-numeric-down"></i> Print Time</a>
+  
+                      <a class="dropdown-item" id="sortPrintTimeUp"><i class="fas fa-sort-numeric-up"></i> Print Time</a>
+                             <div class="dropdown-divider"></div>
+                      <a class="dropdown-item" id="sortDateDown"><i class="fas fa-sort-numeric-down"></i> Upload Date</a>
+  
+                      <a class="dropdown-item" id="sortDateUp"><i class="fas fa-sort-numeric-up"></i> Upload Date</a>
+                          </div>
+                        </div>
+
+                        <label class="btn btn-success float-left mr-1 mb-0 bg-colour-1" for="fileUploadBtn"><i class="fas fa-file-import"></i> Upload File(s)</label>
+                        <input id="fileUploadBtn" multiple accept=".gcode,.gco,.g" type="file" class="btn btn-success float-left bg-colour-1" id="uploadFileBtn">
+                        <label class="btn btn-info float-left mr-1 mb-0 bg-colour-2" for="fileUploadPrintBtn"><i class="fas fa-file-import"></i> Upload and Print</label>
+                        <input id="fileUploadPrintBtn" accept=".gcode,.gco,.g" type="file" class="btn btn-success float-left bg-colour-2" id="uploadFileBtn">
+                      <button
+                        id="createFolderBtn"
+                        type="button"
+                        class="btn btn-warning float-left mr-1 mb-0 bg-colour-3"
+                        data-toggle="collapse"
+                        href="#createFolder"
+                        role="button"
+                        aria-expanded="false"
+                        aria-controls="createFolder"
+                      >
+                        <i class="fas fa-folder-plus"></i> Create Folder
+                      </button>
+                      <button id="fileReSync" type="button" class="btn btn-primary mb-0 bg-colour-4">
+                        <i class="fas fa-sync"></i> Re-Sync
+                      </button>
+                      </div>
+  
+                  </div>
+  
+                      <div id="fileList-${printer._id}" class="list-group" style="height:500px; min-height:500px;max-height:500px; overflow-y:scroll;">
+  
+                      </div>
+              </div>
+          </div>
+              <div class="row">
+                  <div class="col-lg-12 col-xl-6">
+                    <div class="col-12">
+                                      <center>
+                                          <h5>Print Status</h5>
+                                      </center>
+                                      <hr>
+                                  </div>
+                                  <div class="col-12">
+                                                                  <div class="progress mb-2">
+                                    <div id="pmProgress" class="progress-bar" role="progressbar progress-bar-striped" style="width:100%" aria-valuenow="100%" aria-valuemin="0" aria-valuemax="100">Loading...
+                                    </div>
+                                  </div>
+                </div>
+                <div class="row">
+                <div id="fileThumbnail" class="col-12">
+  
+                </div>
+                <div class="col-12">
+                <center>
+                                 <b class="mb-1">File Name: </b><br><p title="Loading..." class="tag mb-1" id="pmFileName">Loading...</p>
+  </center>
+  </div>
+                  <div class="col-12">
+                <center>
+                                 <b id="resentTitle" class="mb-1 d-none">Resend Statistics: </b><br><p title="Current job resend ratio" class="tag mb-1 d-none" id="printerResends">Loading...</p>
+  </center>
+  </div>
+                  <div class="col-lg-12 col-xl-6">
+                     <center>
+                  <b>Expected Completion Date: </b><p class="mb-1" id="pmExpectedCompletionDate">Loading...</p>
+  
+                    <b>Print Time Remaining: </b><p class="mb-1" id="pmTimeRemain">Loading...</p>
+                    <b>Print Time Elapsed: </b><p class="mb-1" id="pmTimeElapsed">Loading...</p>
+                    <b>Current Z: </b><p class="mb-1" id="pmCurrentZ">Loading...</p>
+                    <b>Expected Job Cost: </b><p class="mb-1" id="pmJobCosts">Loading...</p></center>
+              </div>
+                  <div class="col-lg-12 col-xl-6">
+                                 <center>
+                   <b>Expected Print Time: </b><p class="mb-1" id="pmExpectedTime">Loading...</p>
+          <b class="mb-1">Expected Units: </b><br><p class="tag mb-1" id="pmExpectedWeight">Loading...</p>
+          <b class="mb-1">Expected Filament Costs: </b><br><p class="tag mb-1" id="pmExpectedFilamentCost">Loading...</p>
+          <b class="mb-1">Expected Printer Costs: </b><br><p class="tag mb-1" id="pmExpectedPrinterCost">Loading...</p>
+  
+  
+                                 </center>
+  </div>
+                </div>
+                  </div>
+                  <div class="col-lg-12 col-xl-6">
+                 <div class="row">
+                    <div class="col-12">
+                        <center>
+                            <h5>Tools</h5>
+                        </center>
+                        <hr>
+                    </div>
+                </div>
+                <div class="row">
+                  <div class="col-12">
+                      <button id="pmTempTime" type="button" class="btn btn-secondary btn-sm float-right" disabled>Updated: <i class="far fa-clock"></i> Never</button>
+                  </div>
+                </div>
+                <div class="row" id="pmToolTemps">
+  
+                </div>
+                <div class="row">
+                    <div id="pmBedTemp" class="col-lg-6">
+                    
+                    </div>
+                    <div id="pmChamberTemp" class="col-lg-6">
+                    
+                    </div>
+                </div>
+                  </div>
+                </div>
+  
+                       
+                 <div class="row">
+                    <div class="col-12">
+                        <center>
+                            <h5>Terminal</h5>
+                        </center>
+                        <hr>
+                    </div>
+                </div>
+                <div class="row">
+                 <div id="terminal" class="terminal-window bg-secondary">
+                  </div>
+                    <div class="input-group">
+                      <textarea id="terminalInput" type="text" class="form-control" placeholder="" aria-label="" aria-describedby="basic-addon2"></textarea>
+                      <div class="input-group-append">
+                        <button class="btn btn-secondary" id="terminalInputBtn" type="submit">Send</button>
+                      </div>
+                    </div>
+                    <form class="was-validated">
+                      <div class="custom-control custom-checkbox mb-3">
+                        <input 
+                        type="checkbox" class="custom-control-input" id="tempMessages" required checked>
+                        <label class="custom-control-label" for="tempMessages">Temperature Messages</label>
+                        <div class="valid-feedback">Showing temperature messages</div>
+                        <div class="invalid-feedback">Not showing temperature messages</div>
+                      </div>
+                      </form>
+                      <form class="was-validated">
+                       <div class="custom-control custom-checkbox mb-3">
+                        <input
+                        type="checkbox" class="custom-control-input" id="sdMessages" required checked>
+                        <label class="custom-control-label" for="sdMessages">SD Messages</label>
+                        <div class="valid-feedback">Showing sd messages</div>
+                        <div class="invalid-feedback">Not showing sd messages</div>
+                      </div>
+                      </form>
+                      <form class="was-validated">
+                      <div class="custom-control custom-checkbox mb-3">
+                        <input
+                        type="checkbox" class="custom-control-input" id="waitMessages" required checked>
+                        <label class="custom-control-label" for="waitMessages">Wait Responses</label>
+                        <div class="valid-feedback">Showing wait responses</div>
+                        <div class="invalid-feedback">Not showing wait responses</div>
+                      </div>
+                      </form>
+                </div>
+  
+            </div>
+            `;
+    } else {
+      document.getElementById("printerControls").innerHTML = `
           <div class="row">
               <div class="col-lg-3">
              <div class="row">
@@ -601,6 +999,7 @@ export default class PrinterManager {
   
             </div>
             `;
+    }
 
     let camURL = "";
     if (
@@ -1264,13 +1663,18 @@ export default class PrinterManager {
     });
     const submitTerminal = async function (e) {
       let input = elements.terminal.input.value.match(/[^\r\n]+/g);
-      input = input.map(function (name) {
-        if (!name.includes("=")) {
-          return name.toLocaleUpperCase();
-        } else {
-          return name;
-        }
-      });
+      console.log(input);
+      if (input !== null) {
+        input = input.map(function (name) {
+          if (!name.includes("=")) {
+            return name.toLocaleUpperCase();
+          } else {
+            return name;
+          }
+        });
+      } else {
+        return null;
+      }
 
       elements.terminal.input.value = "";
 
@@ -1297,6 +1701,7 @@ export default class PrinterManager {
       if (e.key === "Enter" && !e.shiftKey) {
         submitTerminal(e);
       }
+      console.log(e.key);
     });
     elements.terminal.sendBtn.addEventListener("click", async (e) => {
       submitTerminal(e);
@@ -1559,11 +1964,11 @@ export default class PrinterManager {
         if (printer.currentJob.expectedFilamentCosts !== null) {
           printer.currentJob.expectedFilamentCosts.forEach((unit) => {
             const firstKey = Object.keys(unit)[0];
+            let theLength = parseFloat(unit[firstKey].length);
+            let theWeight = parseFloat(unit[firstKey].weight);
             usageDisplay += `<p class="mb-0"><b>${
               unit[firstKey].toolName
-            }: </b>${unit[firstKey].length.toFixed(2)}m / ${unit[
-              firstKey
-            ].weight.toFixed(2)}g</p>`;
+            }: </b>${theLength.toFixed(2)}m / ${theWeight.toFixed(2)}g</p>`;
           });
 
           filamentCost += `<p class="mb-0"><b>Total: </b>${printer.currentJob.expectedTotals.spoolCost.toFixed(